refactor(core structure): Move files PE-608
