+++ src/ardrive.ts
@@ -18,9 +18,9 @@ import {
 	ArFSDriveTransactionData,
 	ArFSFolderTransactionData
 } from './arfs/arfs_trx_data_types';
-import { ArFSDAO } from './arfsdao';
+import { ArFSDAO } from './arfs/arfsdao';
 import { CommunityOracle } from './community/community_oracle';
-import { deriveDriveKey } from './crypto';
+import { deriveDriveKey } from './utils/crypto';
 import { ARDataPriceEstimator } from './pricing/ar_data_price_estimator';
 import { ARDataPriceRegressionEstimator } from './pricing/ar_data_price_regression_estimator';
 import {
