+++ src/arfs/arfs_builders/arfs_folder_builders.ts
@@ -10,7 +10,7 @@ import {
 	GQLNodeInterface,
 	GQLTagInterface
 } from '../../types';
-import { fileDecrypt } from '../../crypto';
+import { fileDecrypt } from '../../utils/crypto';
 import { Utf8ArrayToStr } from '../../utils/common';
 import { stubEntityID } from '../../utils/stubs';
 import { ArFSPublicFolder, ArFSPrivateFolder } from '../arfs_entities';
