+++ tests/integration/ardrive.int.test.ts
@@ -4,9 +4,9 @@ import { stub } from 'sinon';
 import { ArDrive } from '../../src/ardrive';
 import { RootFolderID } from '../../src/arfs/arfs_builders/arfs_folder_builders';
 import { wrapFileOrFolder, ArFSFileToUpload } from '../../src/arfs/arfs_file_wrapper';
-import { ArFSDAO, PrivateDriveKeyData } from '../../src/arfsdao';
+import { ArFSDAO, PrivateDriveKeyData } from '../../src/arfs/arfsdao';
 import { ArDriveCommunityOracle } from '../../src/community/ardrive_community_oracle';
-import { deriveDriveKey } from '../../src/crypto';
+import { deriveDriveKey } from '../../src/utils/crypto';
 import { ARDataPriceRegressionEstimator } from '../../src/pricing/ar_data_price_regression_estimator';
 import { GatewayOracle } from '../../src/pricing/gateway_oracle';
 import { DriveKey, FeeMultiple, EID, W, UnixTime, ArFSResult, Winston, DrivePrivacy, FileID } from '../../src/types';
